% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_delete_alias}
\alias{gamelift_delete_alias}
\title{Deletes an alias}
\usage{
gamelift_delete_alias(AliasId)
}
\arguments{
\item{AliasId}{[required] Unique identifier for a fleet alias. Specify the alias you want to
delete.}
}
\description{
Deletes an alias. This action removes all record of the alias. Game
clients attempting to access a server process using the deleted alias
receive an error. To delete an alias, specify the alias ID to be
deleted.
}
\details{
\itemize{
\item CreateAlias
\item ListAliases
\item DescribeAlias
\item UpdateAlias
\item DeleteAlias
\item ResolveAlias
}
}
\section{Request syntax}{
\preformatted{gamelift$delete_alias(
  AliasId = "string"
)
}
}

\keyword{internal}
