% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_delete_build}
\alias{gamelift_delete_build}
\title{Deletes a build}
\usage{
gamelift_delete_build(BuildId)
}
\arguments{
\item{BuildId}{[required] Unique identifier for a build to delete.}
}
\description{
Deletes a build. This action permanently deletes the build record and
any uploaded build files.
}
\details{
To delete a build, specify its ID. Deleting a build does not affect the
status of any active fleets using the build, but you can no longer
create new fleets with the deleted build.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/build-intro.html}{Working with Builds}

\strong{Related operations}
\itemize{
\item CreateBuild
\item ListBuilds
\item DescribeBuild
\item UpdateBuild
\item DeleteBuild
}
}
\section{Request syntax}{
\preformatted{gamelift$delete_build(
  BuildId = "string"
)
}
}

\keyword{internal}
