% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_delete_fleet}
\alias{gamelift_delete_fleet}
\title{Deletes everything related to a fleet}
\usage{
gamelift_delete_fleet(FleetId)
}
\arguments{
\item{FleetId}{[required] Unique identifier for a fleet to be deleted.}
}
\description{
Deletes everything related to a fleet. Before deleting a fleet, you must
set the fleet's desired capacity to zero. See UpdateFleetCapacity.
}
\details{
This action removes the fleet's resources and the fleet record. Once a
fleet is deleted, you can no longer use that fleet.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/fleets-intro.html}{Working with Fleets}.

\strong{Related operations}
\itemize{
\item CreateFleet
\item ListFleets
\item DeleteFleet
\item Describe fleets:
\item DescribeFleetAttributes
\item DescribeFleetCapacity
\item DescribeFleetPortSettings
\item DescribeFleetUtilization
\item DescribeRuntimeConfiguration
\item DescribeEC2InstanceLimits
\item DescribeFleetEvents
\item Update fleets:
\item UpdateFleetAttributes
\item UpdateFleetCapacity
\item UpdateFleetPortSettings
\item UpdateRuntimeConfiguration
\item Manage fleet actions:
\item StartFleetActions
\item StopFleetActions
}
}
\section{Request syntax}{
\preformatted{gamelift$delete_fleet(
  FleetId = "string"
)
}
}

\keyword{internal}
