% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_stop_matchmaking}
\alias{gamelift_stop_matchmaking}
\title{Cancels a matchmaking ticket that is currently being processed}
\usage{
gamelift_stop_matchmaking(TicketId)
}
\arguments{
\item{TicketId}{[required] Unique identifier for a matchmaking ticket.}
}
\description{
Cancels a matchmaking ticket that is currently being processed. To stop
the matchmaking operation, specify the ticket ID. If successful, work on
the ticket is stopped, and the ticket status is changed to \code{CANCELLED}.
}
\details{
\itemize{
\item StartMatchmaking
\item DescribeMatchmaking
\item StopMatchmaking
\item AcceptMatch
\item StartMatchBackfill
}
}
\section{Request syntax}{
\preformatted{gamelift$stop_matchmaking(
  TicketId = "string"
)
}
}

\keyword{internal}
